Final Write of the new build script
Move file source file to the src directory.
