+++ WebControllerAddon.py
@@ -1,0 +1,20 @@
+bl_info={'name':'Blender Web Controller','category':'Game Engine'}
+import bpy
+import math
+class BlenderWebController_pl(bpy.types.Panel):
+ bl_idname='game.panel.webcontroller';bl_label='Web Controller';bl_space_type='PROPERTIES';bl_region_type='WINDOW';bl_context='render'
+ def draw(self,context):layout=self.layout;obj=context.object;row=layout.row();row.operator('game.webcontroller')
+class BlenderWebController_op(bpy.types.Operator):
+ 'Launches a website to contoller a new BGE camera';bl_idname='game.webcontroller';bl_label='Setup Web Controller';bl_options={'REGISTER','UNDO'};bl_space_type='PROPERTIES';bl_region_type='WINDOW'
+ @classmethod
+ def poll(self,context):return True
+ def execute(self,context):
+  bpy.context.scene.render.engine='BLENDER_GAME';bpy.context.scene.game_settings.show_debug_properties=True;bpy.ops.text.new();bpy.data.texts[-1].name='StartServer';bpy.data.texts['StartServer'].from_string("from server import WebSocketHttpServer\nfrom handler import BlenderHandler\nimport threading\ndef main():\n if 'Server' in bge.logic.globalDict:return\n else:\n  print('\\n\\n------------------ BUILDING THE SERVER -----------------------');scene=bge.logic.getCurrentScene();scene.active_camera=scene.objects['ControllerView'];cont=bge.logic.getCurrentController();http_addr='',cont.owner['Website Port'];ws_addr='',cont.owner['Socket Port'];bge.logic.globalDict['Server']=WebSocketHttpServer(BlenderHandler,http_address=http_addr,ws_address=ws_addr)\n  if bge.logic.globalDict['Server'].start():cont.owner['Website Address']=bge.logic.globalDict['Server'].httpServer.get_address();cont.owner['Socket Address']=bge.logic.globalDict['Server'].wsServer.get_address();bge.logic.globalDict['Server'].launch_webpage()\n  else:print('Error Starting Server')\nif __name__=='__main__':main()");bpy.ops.text.new();bpy.data.texts[-1].name='EndServer';bpy.data.texts['EndServer'].from_string("import bge\nif 'Server' in bge.logic.globalDict:bge.logic.globalDict['Server'].stop();cont=bge.logic.getCurrentController();cont.activate(cont.actuators['QuitGame'])");bpy.ops.text.new();bpy.data.texts[-1].name='handler.py';bpy.data.texts['handler.py'].use_module=True;bpy.data.texts['handler.py'].from_string("import server\nimport bge\nimport json\nfrom time import sleep\nclass BlenderHandler(server.WebSocketHandler):\n def on_message(self,msg):\n  cont=bge.logic.getCurrentController();msg_info=msg\n  if 'Actuator' in msg_info:\n   direction=msg_info['Actuator'];print('Moving:',direction)\n   if 'Speed' in msg_info:\n    speed=msg_info['Speed'];print('Speed: ',speed);rot=cont.actuators[direction].dRot;loc=cont.actuators[direction].dLoc;print(sum(rot))\n    if sum(rot)!=0:rot=[x/abs(sum(rot)) for x in rot]\n    print(sum(loc))\n    if sum(loc)!=0:loc=[x/abs(sum(loc)) for x in loc]\n    cont.actuators[direction].dLoc=[x*speed for x in loc];cont.actuators[direction].dRot=[x*speed for x in rot]\n   cont.activate(cont.actuators[direction])\n  if 'Stop' in msg_info:\n   for act in cont.actuators:cont.deactivate(act)\n  if 'Reset' in msg_info:\n   try:cont.owner.worldOrientation=[[1.,0.,0.],[0.,1.,0.],[0.,0.,1.]];scene=bge.logic.getCurrentScene();scene.objects['ControllerView'].localPosition=0.,-2.5,0.\n   except e:print('Error: '+e)\n def send_message(self,msg):pass\n def on_close(self):cont=bge.logic.getCurrentController();cont.activate(cont.actuators['SendQuit'])");bpy.ops.text.new();bpy.data.texts[-1].name='server.py';bpy.data.texts['server.py'].use_module=True;bpy.data.texts['server.py'].from_string('import tempfile\nimport http.server\nimport threading\nimport os\nimport socket\nimport socketserver\nimport base64\nimport hashlib\nimport struct\nimport shutil\nimport webbrowser\nimport json\nfrom io import StringIO\nfrom string import Template\ndef writeWebsite(file,new_address):site=Template(\'<!DOCTYPE html>\\n<html lang=en><head><meta charset=utf-8><meta name=viewport content="width=device-width, user-scalable=no, initial-scale=1, height=device-height"><title>BlenderWebController</title><link rel=stylesheet href=http://cdnjs.cloudflare.com/ajax/libs/jquery-mobile/1.4.1/jquery.mobile.min.css><style>/*!* jQuery Mobile 1.4.1 * Git HEAD hash:18c1e32bfc4e0e92756dedc105d799131607f5bb <>Date:Wed Feb 12 2014 22:15:20 UTC * http://jquerymobile.com * * Copyright 2010,2014 jQuery Foundation,Inc. and other contributors * Released under the MIT license. * http://jquery.org/license * */ html{font-size:100%}body,input,select,textarea,button,.ui-btn{font-size:1em;line-height:1.3;font-family:sans-serif}legend,.ui-input-text input,.ui-input-search input{color:inherit;text-shadow:inherit}.ui-mobile label,div.ui-controlgroup-label{font-weight:normal;font-size:16px}.ui-field-contain{border-bottom-color:#828282;border-bottom-color:rgba(0,0,0,.15);border-bottom-width:1px;border-bottom-style:solid}.table-stroke thead th,.table-stripe thead th,.table-stripe tbody tr:last-child{border-bottom:1px solid #d6d6d6;border-bottom:1px solid rgba(0,0,0,.1)}.table-stroke tbody th,.table-stroke tbody td{border-bottom:1px solid #e6e6e6;border-bottom:1px solid rgba(0,0,0,.05)}.table-stripe.table-stroke tbody tr:last-child th,.table-stripe.table-stroke tbody tr:last-child td{border-bottom:0}.table-stripe tbody tr:nth-child(odd) td,.table-stripe tbody tr:nth-child(odd) th{background-color:#eee;background-color:rgba(0,0,0,.04)}.ui-btn,label.ui-btn{font-weight:bold;border-width:1px;border-style:solid}.ui-btn:link{text-decoration:none!important}.ui-btn-active{cursor:pointer}.ui-corner-all{-webkit-border-radius:.6em;border-radius:.6em}.ui-btn-corner-all,.ui-btn.ui-corner-all,.ui-slider-track.ui-corner-all,.ui-flipswitch.ui-corner-all,.ui-li-count{-webkit-border-radius:.3125em;border-radius:.3125em}.ui-btn-icon-notext.ui-btn-corner-all,.ui-btn-icon-notext.ui-corner-all{-webkit-border-radius:1em;border-radius:1em}.ui-btn-corner-all,.ui-corner-all{-webkit-background-clip:padding;background-clip:padding-box}.ui-popup.ui-corner-all>.ui-popup-arrow-guide{left:.6em;right:.6em;top:.6em;bottom:.6em}.ui-shadow{-webkit-box-shadow:0 1px 2px rgba(255,255,255,0.5);-moz-box-shadow:0 1px 2px rgba(255,255,255,0.5);box-shadow:0 1px 2px rgba(255,255,255,0.5)}.ui-shadow-inset{-webkit-box-shadow:inset 0 1px 2px rgba(255,255,255,0.5);-moz-box-shadow:inset 0 1px 2px rgba(255,255,255,0.5);box-shadow:inset 0 1px 2px rgba(255,255,255,0.5)}.ui-overlay-shadow{-webkit-box-shadow:0 0 12px rgba(0,0,0,.6);-moz-box-shadow:0 0 12px rgba(0,0,0,.6);box-shadow:0 0 12px rgba(0,0,0,.6)}.ui-btn-icon-left:after,.ui-btn-icon-right:after,.ui-btn-icon-top:after,.ui-btn-icon-bottom:after,.ui-btn-icon-notext:after{background-color:#000;background-color:rgba(0,0,0,0.7);background-position:center center;background-repeat:no-repeat;-webkit-border-radius:1em;border-radius:1em}.ui-alt-icon.ui-btn:after,.ui-alt-icon .ui-btn:after,html .ui-alt-icon.ui-checkbox-off:after,html .ui-alt-icon.ui-radio-off:after,html .ui-alt-icon .ui-checkbox-off:after,html .ui-alt-icon .ui-radio-off:after{background-color:#000;background-color:rgba(0,0,0,.15)}.ui-nodisc-icon.ui-btn:after,.ui-nodisc-icon .ui-btn:after{background-color:transparent}.ui-shadow-icon.ui-btn:after,.ui-shadow-icon .ui-btn:after{-webkit-box-shadow:0 1px 0 rgba(255,255,255,.4);-moz-box-shadow:0 1px 0 rgba(255,255,255,.4);box-shadow:0 1px 0 rgba(255,255,255,.4)}.ui-btn.ui-checkbox-off:after,.ui-btn.ui-checkbox-on:after,.ui-btn.ui-radio-off:after,.ui-btn.ui-radio-on:after{display:block;width:18px;height:18px;margin:-9px 2px 0 2px}.ui-checkbox-off:after,.ui-btn.ui-radio-off:after{filter:Alpha(Opacity=30);opacity:.3}.ui-btn.ui-checkbox-off:after,.ui-btn.ui-checkbox-on:after{-webkit-border-radius:.1875em;border-radius:.1875em}.ui-radio .ui-btn.ui-radio-on:after{background-image:none;background-color:#fff;width:8px;height:8px;border-width:5px;border-style:solid}.ui-alt-icon.ui-btn.ui-radio-on:after,.ui-alt-icon .ui-btn.ui-radio-on:after{background-color:#000}.ui-icon-loading{background:url("images/ajax-loader.gif");background-size:2.875em 2.875em}.ui-bar-a,.ui-page-theme-a .ui-bar-inherit,html .ui-bar-a .ui-bar-inherit,html .ui-body-a .ui-bar-inherit,html body .ui-group-theme-a .ui-bar-inherit{background:#CDBFAC;border-color:#f95806;color:#F60;text-shadow:0 0 5px #F5EDE3;font-weight:bold}.ui-bar-a{border-width:1px;border-style:solid}.ui-overlay-a,.ui-page-theme-a,.ui-page-theme-a .ui-panel-wrapper{background:#F5EDE3;border-color:#f60;color:#333;text-shadow:0 0 0 #000}.ui-body-a,.ui-page-theme-a .ui-body-inherit,html .ui-bar-a .ui-body-inherit,html .ui-body-a .ui-body-inherit,html body .ui-group-theme-a .ui-body-inherit,html .ui-panel-page-container-a{background:#fff;border-color:#154890;color:#333;text-shadow:0 2px 0 #f3f3f3}.ui-body-a{border-width:1px;border-style:solid}.ui-page-theme-a a,html .ui-bar-a a,html .ui-body-a a,html body .ui-group-theme-a a{color:#69f;font-weight:bold}.ui-page-theme-a a:visited,html .ui-bar-a a:visited,html .ui-body-a a:visited,html body .ui-group-theme-a a:visited{color:#38c}.ui-page-theme-a a:hover,html .ui-bar-a a:hover,html .ui-body-a a:hover,html body .ui-group-theme-a a:hover{color:#059}.ui-page-theme-a a:active,html .ui-bar-a a:active,html .ui-body-a a:active,html body .ui-group-theme-a a:active{color:#059}.ui-page-theme-a .ui-btn,html .ui-bar-a .ui-btn,html .ui-body-a .ui-btn,html body .ui-group-theme-a .ui-btn,html head+body .ui-btn.ui-btn-a,.ui-page-theme-a .ui-btn:visited,html .ui-bar-a .ui-btn:visited,html .ui-body-a .ui-btn:visited,html body .ui-group-theme-a .ui-btn:visited,html head+body .ui-btn.ui-btn-a:visited{background:#E1D4C0;border-color:#154890;color:#333;text-shadow:0 2px 0 #f3f3f3}.ui-page-theme-a .ui-btn:hover,html .ui-bar-a .ui-btn:hover,html .ui-body-a .ui-btn:hover,html body .ui-group-theme-a .ui-btn:hover,html head+body .ui-btn.ui-btn-a:hover{background:#CDBFAC;border-color:#154890;color:#333;text-shadow:0 2px 0 #f3f3f3}.ui-page-theme-a .ui-btn:active,html .ui-bar-a .ui-btn:active,html .ui-body-a .ui-btn:active,html body .ui-group-theme-a .ui-btn:active,html head+body .ui-btn.ui-btn-a:active{background:#69F;border-color:#154890;color:#333;text-shadow:0 2px 0 #f3f3f3}.ui-page-theme-a .ui-btn.ui-btn-active,html .ui-bar-a .ui-btn.ui-btn-active,html .ui-body-a .ui-btn.ui-btn-active,html body .ui-group-theme-a .ui-btn.ui-btn-active,html head+body .ui-btn.ui-btn-a.ui-btn-active,.ui-page-theme-a .ui-checkbox-on:after,html .ui-bar-a .ui-checkbox-on:after,html .ui-body-a .ui-checkbox-on:after,html body .ui-group-theme-a .ui-checkbox-on:after,.ui-btn.ui-checkbox-on.ui-btn-a:after,.ui-page-theme-a .ui-flipswitch-active,html .ui-bar-a .ui-flipswitch-active,html .ui-body-a .ui-flipswitch-active,html body .ui-group-theme-a .ui-flipswitch-active,html body .ui-flipswitch.ui-bar-a.ui-flipswitch-active,.ui-page-theme-a .ui-slider-track .ui-btn-active,html .ui-bar-a .ui-slider-track .ui-btn-active,html .ui-body-a .ui-slider-track .ui-btn-active,html body .ui-group-theme-a .ui-slider-track .ui-btn-active,html body div.ui-slider-track.ui-body-a .ui-btn-active{background-color:#69F;border-color:#154890;color:#fff;text-shadow:1px 1px 1px #333}.ui-page-theme-a .ui-radio-on:after,html .ui-bar-a .ui-radio-on:after,html .ui-body-a .ui-radio-on:after,html body .ui-group-theme-a .ui-radio-on:after,.ui-btn.ui-radio-on.ui-btn-a:after{border-color:#69F}.ui-page-theme-a .ui-btn:focus,html .ui-bar-a .ui-btn:focus,html .ui-body-a .ui-btn:focus,html body .ui-group-theme-a .ui-btn:focus,html head+body .ui-btn.ui-btn-a:focus,.ui-page-theme-a .ui-focus,html .ui-bar-a .ui-focus,html .ui-body-a .ui-focus,html body .ui-group-theme-a .ui-focus,html head+body .ui-btn-a.ui-focus,html head+body .ui-body-a.ui-focus{-webkit-box-shadow:0 0 12px #69F;-moz-box-shadow:0 0 12px #69F;box-shadow:0 0 12px #69F}.ui-bar-b,.ui-page-theme-b .ui-bar-inherit,html .ui-bar-b .ui-bar-inherit,html .ui-body-b .ui-bar-inherit,html body .ui-group-theme-b .ui-bar-inherit{background:#E47297;border-color:#AA0114;color:#AA0114;text-shadow:0 0 5px #fFE9E8;font-weight:bold}.ui-bar-b{border-width:1px;border-style:solid}.ui-overlay-b,.ui-page-theme-b,.ui-page-theme-b .ui-panel-wrapper{background:#292929;border-color:#bbb;color:#292929;text-shadow:0 1px 0 #f3f3f3}.ui-body-b,.ui-page-theme-b .ui-body-inherit,html .ui-bar-b .ui-body-inherit,html .ui-body-b .ui-body-inherit,html body .ui-group-theme-b .ui-body-inherit,html .ui-panel-page-container-b{background:#292929;border-color:#fff;color:#333;text-shadow:0 1px 0 #f3f3f3}.ui-body-b{border-width:1px;border-style:solid}.ui-page-theme-b a,html .ui-bar-b a,html .ui-body-b a,html body .ui-group-theme-b a{color:#38c;font-weight:bold}.ui-page-theme-b a:visited,html .ui-bar-b a:visited,html .ui-body-b a:visited,html body .ui-group-theme-b a:visited{color:#38c}.ui-page-theme-b a:hover,html .ui-bar-b a:hover,html .ui-body-b a:hover,html body .ui-group-theme-b a:hover{color:#059}.ui-page-theme-b a:active,html .ui-bar-b a:active,html .ui-body-b a:active,html body .ui-group-theme-b a:active{color:#059}.ui-page-theme-b .ui-btn,html .ui-bar-b .ui-btn,html .ui-body-b .ui-btn,html body .ui-group-theme-b .ui-btn,html head+body .ui-btn.ui-btn-b,.ui-page-theme-b .ui-btn:visited,html .ui-bar-b .ui-btn:visited,html .ui-body-b .ui-btn:visited,html body .ui-group-theme-b .ui-btn:visited,html head+body .ui-btn.ui-btn-b:visited{background:#f6f6f6;border-color:#ddd;color:#333;text-shadow:0 1px 0 #f3f3f3}.ui-page-theme-b .ui-btn:hover,html .ui-bar-b .ui-btn:hover,html .ui-body-b .ui-btn:hover,html body .ui-group-theme-b .ui-btn:hover,html head+body .ui-btn.ui-btn-b:hover{background:#ededed;border-color:#ddd;color:#333;text-shadow:0 1px 0 #f3f3f3}.ui-page-theme-b .ui-btn:active,html .ui-bar-b .ui-btn:active,html .ui-body-b .ui-btn:active,html body .ui-group-theme-b .ui-btn:active,html head+body .ui-btn.ui-btn-b:active{background:#e8e8e8;border-color:#ddd;color:#333;text-shadow:0 1px 0 #f3f3f3}.ui-page-theme-b .ui-btn.ui-btn-active,html .ui-bar-b .ui-btn.ui-btn-active,html .ui-body-b .ui-btn.ui-btn-active,html body .ui-group-theme-b .ui-btn.ui-btn-active,html head+body .ui-btn.ui-btn-b.ui-btn-active,.ui-page-theme-b .ui-checkbox-on:after,html .ui-bar-b .ui-checkbox-on:after,html .ui-body-b .ui-checkbox-on:after,html body .ui-group-theme-b .ui-checkbox-on:after,.ui-btn.ui-checkbox-on.ui-btn-b:after,.ui-page-theme-b .ui-flipswitch-active,html .ui-bar-b .ui-flipswitch-active,html .ui-body-b .ui-flipswitch-active,html body .ui-group-theme-b .ui-flipswitch-active,html body .ui-flipswitch.ui-bar-b.ui-flipswitch-active,.ui-page-theme-b .ui-slider-track .ui-btn-active,html .ui-bar-b .ui-slider-track .ui-btn-active,html .ui-body-b .ui-slider-track .ui-btn-active,html body .ui-group-theme-b .ui-slider-track .ui-btn-active,html body div.ui-slider-track.ui-body-b .ui-btn-active{background-color:#38c;border-color:#1c4a70;color:#fff;text-shadow:0 1px 0 #059}.ui-page-theme-b .ui-radio-on:after,html .ui-bar-b .ui-radio-on:after,html .ui-body-b .ui-radio-on:after,html body .ui-group-theme-b .ui-radio-on:after,.ui-btn.ui-radio-on.ui-btn-b:after{border-color:#38c}.ui-page-theme-b .ui-btn:focus,html .ui-bar-b .ui-btn:focus,html .ui-body-b .ui-btn:focus,html body .ui-group-theme-b .ui-btn:focus,html head+body .ui-btn.ui-btn-b:focus,.ui-page-theme-b .ui-focus,html .ui-bar-b .ui-focus,html .ui-body-b .ui-focus,html body .ui-group-theme-b .ui-focus,html head+body .ui-btn-b.ui-focus,html head+body .ui-body-b.ui-focus{-webkit-box-shadow:0 0 12px #38c;-moz-box-shadow:0 0 12px #38c;box-shadow:0 0 12px #38c}.ui-disabled,.ui-state-disabled,button[disabled],.ui-select .ui-btn.ui-state-disabled{filter:Alpha(Opacity=30);opacity:.3;cursor:default!important;pointer-events:none}.ui-btn:focus,.ui-btn.ui-focus{outline:0}.ui-noboxshadow .ui-shadow,.ui-noboxshadow .ui-shadow-inset,.ui-noboxshadow .ui-overlay-shadow,.ui-noboxshadow .ui-shadow-icon.ui-btn:after,.ui-noboxshadow .ui-shadow-icon .ui-btn:after,.ui-noboxshadow .ui-focus,.ui-noboxshadow .ui-btn:focus,.ui-noboxshadow input:focus,.ui-noboxshadow .ui-panel{-webkit-box-shadow:none!important;-moz-box-shadow:none!important;box-shadow:none!important}.ui-noboxshadow .ui-btn:focus,.ui-noboxshadow .ui-focus{outline-width:1px;outline-style:auto}</style><style>body{-webkit-touch-callout:none!important;height:100%}a{-webkit-user-select:none!important}.ctrlBtn{padding-top:.7em}#SwipeControl{height:100%}#Page-Swipe .fullHeight{position:absolute;top:40px;right:0;bottom:70px;left:0}.iconButton{-webkit-border-radius:.3125em!important;border-radius:.3125em!important}.ui-footer{position:fixed;bottom:0;left:0;right:0}#popup-strength input{display:none}#popup-strength .ui-slider-track{margin-left:15px}</style><script src=http://cdnjs.cloudflare.com/ajax/libs/jquery/1.11.0/jquery.js></script><script src=http://cdnjs.cloudflare.com/ajax/libs/jquery-mobile/1.4.1/jquery.mobile.min.js></script><script src=http://cdnjs.cloudflare.com/ajax/libs/hammer.js/1.0.6/hammer.min.js></script><script src=http://cdnjs.cloudflare.com/ajax/libs/jquery.qrcode/1.0/jquery.qrcode.min.js></script></head><body><div data-role=panel id=mypanel><ul id=nav-list data-role=listview data-inset=true data-divider-theme=a data-theme=a><li data-role=list-divider>Contoller Mode</li><li><a href=#Page-Buttons>Buttons</a></li><li><a href=#Page-Swipe>Swipe</a></li><li data-role=list-divider>Settings</li><li><a href=#Page-Cxn>Connection</a></li><li><a href=#Page-Info>Info</a></li></ul></div><div data-role=page id=Page-Buttons><div data-role=header><a href=#mypanel class="ui-btn ui-icon-bars ui-corner-all ui-btn-icon-notext iconButton"> &nbsp; </a><h1>Button Control</h1><div class=ui-btn-right><a class="ui-btn ui-icon-recycle ui-corner-all ui-btn-icon-notext iconButton ResetModel"> &nbsp; </a><a class="ui-btn ui-icon-lock ui-corner-all ui-btn-icon-notext iconButton MasterLock"> &nbsp; </a></div></div><div role=main class=ui-content><h2 class="ui-bar ui-bar-b ui-corner-all ErrMsg">Not Connected</h2><div class="ui-grid-a ui-responsive"><div class=ui-block-a><div class=ui-grid-solo><div class=ui-block-a><a href=# id=RotateUp class="ui-btn ui-shadow ui-corner-all ui-btn-icon-top ui-icon-arrow-u ctrlBtn ">&nbsp; </a></div></div><div class=ui-grid-a><div class=ui-block-a><a href=# id=RotateLeft class="ui-btn ui-shadow ui-corner-all ui-btn-icon-top ui-icon-arrow-l ctrlBtn ">&nbsp; </a></div><div class=ui-block-b><a href=# id=RotateRight class="ui-btn ui-shadow ui-corner-all ui-btn-icon-top ui-icon-arrow-r ctrlBtn ">&nbsp; </a></div></div><div class=ui-grid-solo><div class=ui-block-a><a href=# id=RotateDown class="ui-btn ui-shadow ui-corner-all ui-btn-icon-top ui-icon-arrow-d ctrlBtn">&nbsp; </a></div></div></div><div class=ui-block-b><div class=ui-grid-a><div class=ui-block-a><a href=# id=ZoomIn class="ui-btn ui-shadow ui-corner-all ui-btn-icon-top ui-icon-plus ctrlBtn">&nbsp; </a></div><div class=ui-block-b><a href=# id=ZoomOut class="ui-btn ui-shadow ui-corner-all  ui-btn-icon-top ui-icon-minus ctrlBtn ">&nbsp; </a></div></div><div class=ui-grid-a><div class=ui-block-a><a href=# id=ZRotateLeft class="ui-btn ui-shadow ui-corner-all ui-btn-icon-top ui-icon-back ctrlBtn ">&nbsp; </a></div><div class=ui-block-b><a href=# id=ZRotateRight class="ui-btn ui-shadow ui-corner-all  ui-btn-icon-top ui-icon-forward ctrlBtn ">&nbsp; </a></div></div></div></div></div><div data-role=footer data-id=main-footer id=button-footer><div data-role=controlgroup data-type=horizontal data-mini=true><a class="ui-btn ui-corner-all ui-shadow "><<</a><a class="ui-btn ui-corner-all ui-shadow "><</a><a class="ui-btn ui-corner-all ui-shadow ">Play</a><a class="ui-btn ui-corner-all ui-shadow ">></a><a class="ui-btn ui-corner-all ui-shadow ">>></a></div><a href=#button-popup-strength data-rel=popup data-position-to=#button-footer class="ui-btn ui-btn-right ui-corner-all ui-shadow ui-icon-gear ui-btn-icon-notext iconButton">Sensitivity</a></div><div data-role=popup id=button-popup-strength class="ui-corner-all ui-content"><a href=# data-rel=back class="ui-btn ui-corner-all ui-shadow ui-btn-a ui-icon-delete ui-btn-icon-notext ui-btn-right">Close</a><h3>Adjust Movement Sensitivity</h3><form><label for=strength class=ui-hidden-accessible>Sensitivity</label><input name=strength id=btn_speed min=1 max=100 value=50 type=range data-highlight=true data-popup-enabled=true></form></div></div><div data-role=page id=Page-Swipe><div data-role=header><a href=#mypanel class="ui-btn ui-icon-bars ui-corner-all ui-btn-icon-notext iconButton"> &nbsp; </a><h1>Swipe Control</h1><div class=ui-btn-right><a class="ui-btn ui-icon-recycle ui-corner-all ui-btn-icon-notext iconButton ResetModel"> &nbsp; </a><a class="ui-btn ui-icon-lock ui-corner-all ui-btn-icon-notext iconButton MasterLock"> &nbsp; </a></div></div><div role=main class="ui-content fullHeight"><h2 class="ui-bar ui-bar-b ui-corner-all ErrMsg">Not Connected</h2><div id=SwipeControl class="ui-body ui-body-a ui-corner-all"><p id=SwipeEvent>Swipe Controls <p><ul><li><b>TAP:</b> Stop </li><li><b>SWIPE:</b> Rotate</li><li><b>PINCH:</b> Zoom </li><li><b>ROTATE:</b> Rotate Z Axis </li></ul></div></div><div data-role=footer data-id=main-footer id=slide-footer><div data-role=controlgroup data-type=horizontal data-mini=true><a class="ui-btn ui-corner-all ui-shadow "><<</a><a class="ui-btn ui-corner-all ui-shadow "><</a><a class="ui-btn ui-corner-all ui-shadow ">Play</a><a class="ui-btn ui-corner-all ui-shadow ">></a><a class="ui-btn ui-corner-all ui-shadow ">>></a></div><a href=#swipe-popup-strength data-rel=popup data-position-to=#slide-footer class="ui-btn ui-btn-right ui-corner-all ui-shadow ui-icon-gear ui-btn-icon-notext iconButton">Sensitivity</a></div><div data-role=popup id=swipe-popup-strength class="ui-corner-all ui-content"><a href=# data-rel=back class="ui-btn ui-corner-all ui-shadow ui-btn-a ui-icon-delete ui-btn-icon-notext ui-btn-right">Close</a><h3>Adjust Movement Sensitivity</h3><form><label for=strength class=ui-hidden-accessible>Sensitivity</label><input name=strength id=swipe_speed min=1 max=100 value=5 type=range data-highlight=true data-popup-enabled=true></form></div></div><div data-role=page id=Page-Cxn><div data-role=header><a href=#mypanel class="ui-btn ui-icon-bars ui-corner-all ui-btn-icon-notext iconButton"> &nbsp; </a><h1>Connection Settings</h1><a href=#popupLogin data-rel=popup data-position-to=window class="ui-btn ui-corner-all ui-shadow ui-icon-camera ui-btn-icon-notext iconButton">&nbsp; </a></div><div role=main class=ui-content><h2 class="ui-bar ui-bar-b ui-corner-all ErrMsg">Not Connected</h2><h3 class="ui-bar ui-bar-a ui-corner-all">Http & Websocket Address</h3><div data-role=controlgroup><input name=httpAddr type=url id=CxnHTTPAddress data-wrapper-class="controlgroup-textinput ui-btn"><input name=wsAddr type=url id=CxnWSAddress data-wrapper-class="controlgroup-textinput ui-btn" data-clear-btn=true><button id=ToggleCxnBtn>Connect</button></div></div><div data-role=popup id=popupLogin data-theme=a class=ui-corner-all><form><div style="padding:10px 20px;"><a href=# data-rel=back class="ui-btn ui-corner-all ui-shadow ui-btn-a ui-icon-delete ui-btn-icon-notext ui-btn-right">Close</a><h3>Scan to connect</h3><span id=CxnQR></span></div></form></div></div><div data-role=page id=Page-Info><div data-role=header><a href=#mypanel class="ui-btn ui-icon-bars ui-corner-all ui-btn-icon-notext iconButton"> &nbsp; </a><h1>Swipe Control</h1><button id=DebugMsgListBtn class="ui-btn ui-icon-check ui-corner-all ui-btn-icon-notext iconButton"> &nbsp; </button></div><div role=main class=ui-content><h2 class="ui-bar ui-bar-b ui-corner-all ErrMsg">Not Connected</h2><h3 class="ui-bar ui-bar-a ui-corner-all">Received Messages</h3><ul data-role=listview data-inset=True id=DebugMsgList></ul></div></div></body><script>var address = "$address";</script><script>var a=null;var b=0;var c=false;var d=/(\\\\\\d{1,3}\\.\\\\\\d{1,3}\\.\\\\\\d{1,3}\\.\\\\\\d{1,3}):(\\\\\\d{1,5})/;window.onbeforeunload=g;function e(a,b){if(typeof b===\\\'undefined\\\')b="Message";$("#DebugMsgList").prepend(\\\'<li><h3>\\\'+b+\\\'</h3><p>\\\'+a+\\\'</p></li>\\\').listview("refresh");}function f(a){$("#ToggleCxnBtn").text("Disconnect");e("Connected!");$(".ErrMsg").fadeOut();};function g(a){e("Connection Closed!");$("#ToggleCxnBtn").text("Connect");$(".ErrMsg").text("Not Connected").fadeIn();};function h(a){msg_data=JSON.parse(a.data);if(msg_data.hasOwnProperty("SLAVE"))m(msg_data.SLAVE);if(msg_data.hasOwnProperty("MASTER_STATUS")){c=msg_data.MASTER_STATUS;n();}e(a.data);};function i(a){e("Error: "+a);$("#ToggleCxnBtn").text("Connect");$(".ErrMsg").text("Error! - Check Msg").fadeIn();};function j(b,c,d){if(typeof d===\\\'undefined\\\')d=0.001;var f={};f[b]=c;f.Speed=d;var g=JSON.stringify(f);if(a)a.send(g);else e("Event: "+g);}function k(){if(a){a.close(1000,"Try to Close");a=null;}else try{address=$("#CxnWSAddress").val();if(d.test(address)){a=new WebSocket(address);a.onopen=f;a.onclose=g;a.onmessage=h;a.onerror=i;}}catch(b){i("Could Not Connected");}}function l(){j("MASTER_REQUEST",!c);}function m(a){if(a){$(".ctrlBtn").addClass(\\\'ui-disabled\\\');$(".ErrMsg").text("Locked").fadeIn();}else{$(".ctrlBtn").removeClass(\\\'ui-disabled\\\');$(".ErrMsg").text("Locked").fadeOut();}}function n(){if(c){e("You are now Master");$(".MasterLock").addClass("danger");}else $(".MasterLock").removeClass("danger");}$(document).ready(function(){$("#CxnHTTPAddress").val(document.URL);$("#CxnWSAddress").val(address);$("#DebugMsgList").listview({create:function(a,b){}});k();$("#CxnQR").qrcode(document.URL);$("#ToggleCxnBtn").click(k);$(".MasterLock").click(l);$("#DebugMsgListBtn").click(function(){$("#DebugMsgList").empty();});$("a").bind("contextmenu",function(a){a.preventDefault();});$("button").bind("contextmenu",function(a){a.preventDefault();});$(".ctrlBtn").on("vmousedown ",function(a){a.preventDefault();j("Actuator",$(this).attr("id"),$("#btn_speed").val()/1000);}).on("vmouseup ",function(){j("Stop","All");});$(".ResetModel").mousedown(function(){j("Reset",true);}).mouseup(function(){j("Stop","All");});var a=document.getElementById(\\\'SwipeControl\\\');var c=Hammer(a,{prevent_default:true,no_mouseevents:true,transform_always_block:true}).on("tap",function(a){j("Stop","All");}).on("release",function(a){j("Stop","All");}).on("dragleft",function(a){j("Actuator","RotateLeft",($("#swipe_speed").val()/1000));a.gesture.stopDetect();}).on("dragright",function(a){j("Actuator","RotateRight",($("#swipe_speed").val()/1000));a.gesture.stopDetect();}).on("dragdown",function(a){j("Actuator","RotateDown",($("#swipe_speed").val()/1000));a.gesture.stopDetect();}).on("dragup",function(a){j("Actuator","RotateUp",($("#swipe_speed").val()/1000));a.gesture.stopDetect();}).on("pinchin",function(a){j("Actuator","ZoomIn",($("#swipe_speed").val()/1000));a.gesture.stopDetect();}).on("pinchout",function(a){j("Actuator","ZoomOut",($("#swipe_speed").val()/1000));a.gesture.stopDetect();}).on("rotate",function(a){if(a.gesture.rotation<0)j("Actuator","ZRotateLeft",($("#swipe_speed").val()/1000));else j("Actuator","ZRotateRight",($("#swipe_speed").val()/1000));a.gesture.stopDetect();});$(document).keydown(function(a){if(a.which==37&&!b){a.preventDefault();b=true;a.shiftKey?j("Actuator","ZRotateLeft"):j("Actuator","RotateLeft");}if(a.which==38&&!b){a.preventDefault();b=true;a.shiftKey?j("Actuator","ZoomOut"):j("Actuator","RotateUp");}if(a.which==39&&!b){a.preventDefault();b=true;a.shiftKey?j("Actuator","ZRotateRight"):j("Actuator","RotateRight");}if(a.which==40&&!b){a.preventDefault();b=true;a.shiftKey?j("Actuator","ZoomIn"):j("Actuator","RotateDown");}if(a.which==27&&!b){b=true;k();}});$(document).keyup(function(a){j("Stop","All");b=false;});});</script><script>$(function(){$("#mypanel").panel().enhanceWithin();});</script></html>\').safe_substitute(address=new_address);html=open(file,\'w\');html.write(site);html.close()\nclass QuiteCGIHandler(http.server.CGIHTTPRequestHandler):\n def log_message(self,format,*args):pass\nclass WebSocketHandler(socketserver.BaseRequestHandler):\n def on_message(self,msg):print(msg)\n def handle_message(self,msg):\n  if self._hasLock():\n   msg_data=json.loads(msg)\n   if \'MASTER_REQUEST\' in msg_data:\n    if msg_data[\'MASTER_REQUEST\']:WebSocketHandler.lock_id=threading.current_thread().ident;self.send_json(dict(MASTER_STATUS=True));print(\'Locking to thread: \',WebSocketHandler.lock_id,\'   :   \',self.id);self.broadcast_all(dict(SLAVE=True))\n    else:WebSocketHandler.lock_id=None;self.send_json(dict(MASTER_STATUS=False));self.broadcast_all(dict(SLAVE=False))\n   self.on_message(msg_data)\n  else:self.send_json(dict(SLAVE=True));print(\'Locked, trashing: \',msg)\n def on_close(self):print(\'Server: Closing Connection for \',self.client_address);self.send_message(\'Server: Closing Connection\')\n def send_message(self,message):print(\'Sending: \',message);self.send_json(dict(MESSAGE=message))\n def send_json(self,data):self.request.sendall(self._pack(json.dumps(data)))\n def broadcast_all(self,data):\n  for t in WebSocketHandler.connections:\n   if t.id==self.id:continue\n   t.send_json(data)\n magic=b\'258EAFA5-E914-47DA-95CA-C5AB0DC85B11\';lock_id=None;connections=[]\n def _hasLock(self):return not WebSocketHandler.lock_id or WebSocketHandler.lock_id==self.id\n def setup(self):print(\'\\nConnection Established\',self.client_address);self.closeHandle=False;self.id=threading.current_thread().ident;self.alive=threading.Event();self.alive.set();WebSocketHandler.connections.append(self)\n def handle(self):\n  try:self.handshake()\n  except:print(\'HANDSHAKE ERROR! - Try using FireFox\')\n def run(self):\n  while self.alive.isSet():\n   msg=self.request.recv(2)\n   if not msg or self.closeHandle or msg[0]==136:print(\'Received Closed\');break\n   length=msg[1]&127\n   if length==126:length=struct.unpack(\'>H\',self.request.recv(2))[0]\n   elif length==127:length=struct.unpack(\'>Q\',self.request.recv(8))[0]\n   masks=self.request.recv(4);decoded=\'\'\n   for char in self.request.recv(length):decoded+=chr(char^masks[len(decoded)%4])\n   self.handle_message(decoded)\n  self.close()\n def close(self,message=\'Cxn Closed\'):self.closeHandle=True;self.request.sendall(self._pack(message,True));self.on_close()\n def handshake(self):\n  key=None;data=self.request.recv(1024).strip()\n  for line in data.splitlines():\n   if b\'Upgrade:\' in line:\n    upgrade=line.split(b\': \')[1]\n    if not upgrade==b\'websocket\':raise Exception(\'Upgrade is Not a websocket!\',data)\n   if b\'Sec-WebSocket-Key:\' in line:key=line.split(b\': \')[1];break\n  if key is None:raise Exception("Couldn\'t find the key?:",data)\n  print(\'Handshaking...   \',end=\'\');digest=self._websocketHash(key);response=\'HTTP/1.1 101 Switching Protocols\\n\';response+=\'Upgrade: websocket\\n\';response+=\'Connection: Upgrade\\n\';response+=\'Sec-WebSocket-Accept: %s\\n\\n\'%digest;self.handshake_done=self.request.send(response.encode());print(\'Sending Connected Message...   \',end=\'\')\n  if self.handshake_done:self.send_message(\'Connected!\')\n  print(\'Connected!\\n\')\n def _websocketHash(self,key):result_string=key+self.magic;sha1_digest=hashlib.sha1(result_string).digest();response_data=base64.encodestring(sha1_digest).strip();response_string=response_data.decode(\'utf8\');return response_string\n def _get_framehead(self,close=False):\n  frame_head=bytearray(2);frame_head[0]=frame_head[0]|1<<7\n  if close:frame_head[0]=frame_head[0]|8<<0\n  else:frame_head[0]=frame_head[0]|1<<0\n  return frame_head\n def _pack(self,data,close=False):\n  frame_head=self._get_framehead(close)\n  if len(data)<126:frame_head[1]=len(data)\n  elif len(data)<2**16-1:frame_head[1]=126;frame_head+=int_to_bytes(len(data),2)\n  elif len(data)<2**64-1:frame_head[1]=127;frame_head+=int_to_bytes(len(data),8)\n  frame=frame_head+data.encode(\'utf-8\');return frame\nclass HTTPServer(threading.Thread):\n def __init__(self,address_info=(\'\',0)):threading.Thread.__init__(self);self.httpd=None;self._start_server(address_info)\n def _start_server(self,address_info):\n  try:self.httpd=http.server.HTTPServer(address_info,QuiteCGIHandler);print(\'HTTP Server on : \',self.get_address())\n  except Exception as e:print(\'The HTTP server could not be started\')\n def get_address(self):\n  if self.httpd is not None:return \'http://{host}:{port}/\'.format(host=socket.gethostbyname(self.httpd.server_name),port=self.httpd.server_port)\n  else:return None\n def run(self):\n  if self.httpd is not None:self.httpd.serve_forever()\n  else:print(\'Error! - HTTP Server is NULL\')\n def stop(self):\n  print(\'Killing Http Server ...\')\n  if self.httpd is not None:self.httpd.shutdown()\n  print(\'Done\')\nclass WebSocketTCPServer(socketserver.ThreadingMixIn,http.server.HTTPServer):\n def __init__(self,server_address,RequestHandlerClass,bind_and_activate=True):socketserver.TCPServer.__init__(self,server_address,RequestHandlerClass,bind_and_activate);self.handlers=[];self.daemon_threads=True\n def finish_request(self,request,client_address):print(\'launching a new request\');t=self.RequestHandlerClass(request,client_address,self);print(\'Request:\',t);self.handlers.append(t);print(\'Num request:\',len(self.handlers));t.run()\n def get_handlers(self):return self.handlers\nclass WebsocketServer(threading.Thread):\n def __init__(self,handler,address_info=(\'\',0)):threading.Thread.__init__(self);self.wsd=None;self.handler=handler;self._start_server(address_info)\n def _start_server(self,address_info):\n  try:self.wsd=WebSocketTCPServer(address_info,self.handler);print(self.get_address())\n  except Exception as e:print(\'Error! - Websocket Server Not Started!\',e)\n def get_address(self):\n  if self.wsd is not None:return \'ws://{host}:{port}/\'.format(host=socket.gethostbyname(self.wsd.server_name),port=self.wsd.server_port)\n  else:return None\n def run(self):\n  if self.wsd is not None:self.wsd.serve_forever()\n  else:print(\'The WebSocket Server is NULL\')\n def stop(self):\n  print(\'Killing WebSocket Server ...\')\n  if self.wsd is not None:\n   for h in self.wsd.handlers:h.alive.clear()\n   self.wsd.shutdown()\n  print(\'Done\')\n def get_handlers(self):return self.wsd.get_handlers()\n def send(self,msg):\n  for h in self.wsd.get_handlers():h.send_message(msg)\nclass WebSocketHttpServer:\n def __init__(self,handler_class=WebSocketHandler,http_address=(\'\',0),ws_address=(\'\',0)):self.http_address=http_address;self.ws_address=ws_address;self.handler=handler_class;self.httpServer=None;self.wsServer=None\n def _clean_server_temp_dir(self):os.chdir(self.cwd);shutil.rmtree(self.tempdir)\n def _make_server_temp_dir(self):self.cwd=os.path.dirname(os.path.realpath(__file__));self.tempdir=tempfile.mkdtemp();os.chdir(self.tempdir);print(\'New temp dir:\',self.tempdir)\n def _make_webpage(self):writeWebsite(self.tempdir+\'/index.html\',self.wsServer.get_address())\n def stop(self):\n  try:self.httpServer.stop();self.wsServer.stop();self._clean_server_temp_dir()\n  except Exception as e:print(\'The Servers were never started\',e)\n def start(self):\n  try:\n   self._make_server_temp_dir();self.httpServer=HTTPServer(self.http_address);self.wsServer=WebsocketServer(self.handler,self.ws_address)\n   if self.wsServer is not None and self.httpServer is not None:self.httpServer.start();self.wsServer.start();self._make_webpage();return True\n   else:print(\'Error Starting The Servers, Something is not Initialized!\');return False\n  except Exception as e:print();print(\'Error!!!, There is some error!\');print(e);print();return False\n def send(self,msg):self.wsServer.send(msg)\n def launch_webpage(self):webbrowser.open(self.httpServer.get_address()+\'index.html\')\n def status(self):\n  if self.wsServer is None or self.httpServer is None:return False\n  else:return True\nif __name__==\'__main__\':print(\'No Main Program!\')');bpy.ops.object.add(type='EMPTY');bpy.context.active_object.name='Controller';bpy.ops.object.game_property_new(type='INT',name='Website Port');bpy.ops.object.game_property_new(type='INT',name='Socket Port');bpy.ops.object.game_property_new(type='STRING',name='Website Address');bpy.ops.object.game_property_new(type='STRING',name='Socket Address');bpy.data.objects['Controller'].location=0.,0.,0.;bpy.data.objects['Controller'].game.properties['Website Port'].value=8000;bpy.data.objects['Controller'].game.properties['Socket Port'].value=0;bpy.data.objects['Controller'].game.properties['Website Address'].show_debug=True;bpy.data.objects['Controller'].game.properties['Socket Address'].show_debug=True;bpy.ops.logic.sensor_add(type='DELAY',name='StartServer');bpy.ops.logic.sensor_add(type='KEYBOARD',name='QuitKey');bpy.ops.logic.sensor_add(type='MESSAGE',name='QuitMsg');bpy.ops.logic.controller_add(type='PYTHON',name='Sever');bpy.ops.logic.controller_add(type='PYTHON',name='QuitSever');bpy.ops.logic.actuator_add(type='MOTION',name='RotateRight');bpy.ops.logic.actuator_add(type='MOTION',name='RotateLeft');bpy.ops.logic.actuator_add(type='MOTION',name='RotateUp');bpy.ops.logic.actuator_add(type='MOTION',name='RotateDown');bpy.ops.logic.actuator_add(type='MOTION',name='ZRotateLeft');bpy.ops.logic.actuator_add(type='MOTION',name='ZRotateRight');bpy.ops.logic.actuator_add(type='MESSAGE',name='SendQuit');bpy.ops.logic.actuator_add(type='GAME',name='QuitGame');bpy.ops.object.add(type='CAMERA');bpy.context.active_object.name='ControllerView';bpy.data.objects['ControllerView'].location=0.,-5.,0.;bpy.data.objects['ControllerView'].rotation_euler=math.radians(90),0.,0.;bpy.data.objects['ControllerView'].parent=bpy.data.objects['Controller'];bpy.ops.logic.actuator_add(type='MOTION',name='ZoomOut');bpy.ops.logic.actuator_add(type='MOTION',name='ZoomIn');con=bpy.data.objects['ControllerView'].constraints.new('DAMPED_TRACK');con.name='ViewLock';con.target=bpy.data.objects['Controller'];con.track_axis='TRACK_NEGATIVE_Z';bpy.context.scene.camera=bpy.data.objects['ControllerView'];bpy.data.objects['Controller'].game.sensors['StartServer'].use_repeat=True;bpy.data.objects['Controller'].game.sensors['QuitKey'].key='Q';bpy.data.objects['Controller'].game.sensors['QuitMsg'].subject='QUIT';bpy.data.objects['Controller'].game.controllers['Sever'].use_priority=True;bpy.data.objects['Controller'].game.controllers['Sever'].text=bpy.data.texts['StartServer'];bpy.data.objects['Controller'].game.controllers['QuitSever'].text=bpy.data.texts['EndServer'];bpy.data.objects['Controller'].game.actuators['QuitGame'].mode='QUIT';bpy.data.objects['Controller'].game.actuators['RotateRight'].offset_rotation=0.,0.,math.radians(-1);bpy.data.objects['Controller'].game.actuators['RotateLeft'].offset_rotation=0.,0.,math.radians(1);bpy.data.objects['Controller'].game.actuators['RotateUp'].offset_rotation=math.radians(1),0.,0.;bpy.data.objects['Controller'].game.actuators['RotateDown'].offset_rotation=math.radians(-1),0.,0.;bpy.data.objects['Controller'].game.actuators['ZRotateLeft'].offset_rotation=0.,math.radians(1),0.;bpy.data.objects['Controller'].game.actuators['ZRotateRight'].offset_rotation=0.,math.radians(-1),0.;bpy.data.objects['ControllerView'].game.actuators['ZoomIn'].offset_location=0.,0.,.01;bpy.data.objects['ControllerView'].game.actuators['ZoomOut'].offset_location=0.,0.,-.01;bpy.data.objects['Controller'].game.actuators['SendQuit'].to_property=bpy.data.objects['Controller'].name;c=bpy.data.objects['Controller'].game.controllers['Sever'];d=bpy.data.objects['Controller'].game.controllers['QuitSever'];bpy.data.objects['Controller'].game.sensors['StartServer'].link(c);bpy.data.objects['Controller'].game.actuators['RotateRight'].link(c);bpy.data.objects['Controller'].game.actuators['RotateLeft'].link(c);bpy.data.objects['Controller'].game.actuators['RotateUp'].link(c);bpy.data.objects['Controller'].game.actuators['RotateDown'].link(c);bpy.data.objects['Controller'].game.actuators['ZRotateLeft'].link(c);bpy.data.objects['Controller'].game.actuators['ZRotateRight'].link(c);bpy.data.objects['Controller'].game.actuators['SendQuit'].link(c);bpy.data.objects['ControllerView'].game.actuators['ZoomIn'].link(c);bpy.data.objects['ControllerView'].game.actuators['ZoomOut'].link(c);bpy.data.objects['Controller'].game.sensors['QuitKey'].link(d);bpy.data.objects['Controller'].game.sensors['QuitMsg'].link(d);bpy.data.objects['Controller'].game.actuators['QuitGame'].link(d)
+  for s in bpy.data.objects['ControllerView'].game.sensors:s.show_expanded=False
+  for s in bpy.data.objects['ControllerView'].game.actuators:s.show_expanded=False
+  for s in bpy.data.objects['Controller'].game.sensors:s.show_expanded=False
+  for s in bpy.data.objects['Controller'].game.actuators:s.show_expanded=False
+  bpy.ops.object.select_all(action='DESELECT');bpy.data.objects['Controller'].select=True;return {'FINISHED'}
+def register():bpy.utils.register_class(BlenderWebController_op);bpy.utils.register_class(BlenderWebController_pl)
+def unregister():bpy.utils.unregister_class(BlenderWebController_op);bpy.utils.unregister_class(BlenderWebController_pl)
+if __name__=='__main__':register()
